Revert "add cuequivariance support"
