+++ mace/__version__.py
@@ -1,3 +1,3 @@
-__version__ = "0.3.8"
+__version__ = "0.3.7"
 
 __all__ = ["__version__"]
