+++ mace/cli/fine_tuning_select.py
@@ -20,9 +20,7 @@ except ImportError:
 
 
 def parse_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser(
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-    )
+    parser = argparse.ArgumentParser()
     parser.add_argument(
         "--configs_pt",
         help="path to XYZ configurations for the pretraining",
