+++ mace/modules/blocks.py
@@ -12,13 +12,6 @@ import torch.nn.functional
 from e3nn import nn, o3
 from e3nn.util.jit import compile_mode
 
-from mace.modules.wrapper_ops import (
-    CuEquivarianceConfig,
-    FullyConnectedTensorProduct,
-    Linear,
-    SymmetricContractionWrapper,
-    TensorProduct,
-)
 from mace.tools.compile import simplify_if_compile
 from mace.tools.scatter import scatter_sum
 
@@ -36,6 +29,7 @@ from .radial import (
     PolynomialCutoff,
     SoftTransform,
 )
+from .symmetric_contraction import SymmetricContraction
 
 
 @compile_mode("script")
@@ -80,16 +74,9 @@ class NonLinearReadoutBlockPartial(torch.nn.Module):
 
 @compile_mode("script")
 class LinearNodeEmbeddingBlock(torch.nn.Module):
-    def __init__(
-        self,
-        irreps_in: o3.Irreps,
-        irreps_out: o3.Irreps,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
+    def __init__(self, irreps_in: o3.Irreps, irreps_out: o3.Irreps):
         super().__init__()
-        self.linear = Linear(
-            irreps_in=irreps_in, irreps_out=irreps_out, cueq_config=cueq_config
-        )
+        self.linear = o3.Linear(irreps_in=irreps_in, irreps_out=irreps_out)
 
     def forward(
         self,
@@ -100,16 +87,9 @@ class LinearNodeEmbeddingBlock(torch.nn.Module):
 
 @compile_mode("script")
 class LinearReadoutBlock(torch.nn.Module):
-    def __init__(
-        self,
-        irreps_in: o3.Irreps,
-        irrep_out: o3.Irreps = o3.Irreps("0e"),
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
+    def __init__(self, irreps_in: o3.Irreps, irrep_out: o3.Irreps = o3.Irreps("0e")):
         super().__init__()
-        self.linear = Linear(
-            irreps_in=irreps_in, irreps_out=irrep_out, cueq_config=cueq_config
-        )
+        self.linear = o3.Linear(irreps_in=irreps_in, irreps_out=irrep_out)
 
     def forward(
         self,
@@ -129,18 +109,13 @@ class NonLinearReadoutBlock(torch.nn.Module):
         gate: Optional[Callable],
         irrep_out: o3.Irreps = o3.Irreps("0e"),
         num_heads: int = 1,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
     ):
         super().__init__()
         self.hidden_irreps = MLP_irreps
         self.num_heads = num_heads
-        self.linear_1 = Linear(
-            irreps_in=irreps_in, irreps_out=self.hidden_irreps, cueq_config=cueq_config
-        )
+        self.linear_1 = o3.Linear(irreps_in=irreps_in, irreps_out=self.hidden_irreps)
         self.non_linearity = nn.Activation(irreps_in=self.hidden_irreps, acts=[gate])
-        self.linear_2 = Linear(
-            irreps_in=self.hidden_irreps, irreps_out=irrep_out, cueq_config=cueq_config
-        )
+        self.linear_2 = o3.Linear(irreps_in=self.hidden_irreps, irreps_out=irrep_out)
 
     def forward(
         self, x: torch.Tensor, heads: Optional[torch.Tensor] = None
@@ -154,20 +129,13 @@ class NonLinearReadoutBlock(torch.nn.Module):
 
 @compile_mode("script")
 class LinearDipoleReadoutBlock(torch.nn.Module):
-    def __init__(
-        self,
-        irreps_in: o3.Irreps,
-        dipole_only: bool = False,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
+    def __init__(self, irreps_in: o3.Irreps, dipole_only: bool = False):
         super().__init__()
         if dipole_only:
             self.irreps_out = o3.Irreps("1x1o")
         else:
             self.irreps_out = o3.Irreps("1x0e + 1x1o")
-        self.linear = Linear(
-            irreps_in=irreps_in, irreps_out=self.irreps_out, cueq_config=cueq_config
-        )
+        self.linear = o3.Linear(irreps_in=irreps_in, irreps_out=self.irreps_out)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:  # [n_nodes, irreps]  # [..., ]
         return self.linear(x)  # [n_nodes, 1]
@@ -181,7 +149,6 @@ class NonLinearDipoleReadoutBlock(torch.nn.Module):
         MLP_irreps: o3.Irreps,
         gate: Callable,
         dipole_only: bool = False,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
     ):
         super().__init__()
         self.hidden_irreps = MLP_irreps
@@ -204,13 +171,9 @@ class NonLinearDipoleReadoutBlock(torch.nn.Module):
             irreps_gated=irreps_gated,
         )
         self.irreps_nonlin = self.equivariant_nonlin.irreps_in.simplify()
-        self.linear_1 = Linear(
-            irreps_in=irreps_in, irreps_out=self.irreps_nonlin, cueq_config=cueq_config
-        )
-        self.linear_2 = Linear(
-            irreps_in=self.hidden_irreps,
-            irreps_out=self.irreps_out,
-            cueq_config=cueq_config,
+        self.linear_1 = o3.Linear(irreps_in=irreps_in, irreps_out=self.irreps_nonlin)
+        self.linear_2 = o3.Linear(
+            irreps_in=self.hidden_irreps, irreps_out=self.irreps_out
         )
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:  # [n_nodes, irreps]  # [..., ]
@@ -295,25 +258,22 @@ class EquivariantProductBasisBlock(torch.nn.Module):
         correlation: int,
         use_sc: bool = True,
         num_elements: Optional[int] = None,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
     ) -> None:
         super().__init__()
 
         self.use_sc = use_sc
-        self.symmetric_contractions = SymmetricContractionWrapper(
+        self.symmetric_contractions = SymmetricContraction(
             irreps_in=node_feats_irreps,
             irreps_out=target_irreps,
             correlation=correlation,
             num_elements=num_elements,
-            cueq_config=cueq_config,
         )
         # Update linear
-        self.linear = Linear(
+        self.linear = o3.Linear(
             target_irreps,
             target_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=cueq_config,
         )
 
     def forward(
@@ -340,7 +300,6 @@ class InteractionBlock(torch.nn.Module):
         hidden_irreps: o3.Irreps,
         avg_num_neighbors: float,
         radial_MLP: Optional[List[int]] = None,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
     ) -> None:
         super().__init__()
         self.node_attrs_irreps = node_attrs_irreps
@@ -353,7 +312,6 @@ class InteractionBlock(torch.nn.Module):
         if radial_MLP is None:
             radial_MLP = [64, 64, 64]
         self.radial_MLP = radial_MLP
-        self.cueq_config = cueq_config
 
         self._setup()
 
@@ -407,29 +365,23 @@ class TensorProductWeightsBlock(torch.nn.Module):
 @compile_mode("script")
 class ResidualElementDependentInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
-
-        # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
             self.node_feats_irreps, self.edge_attrs_irreps, self.target_irreps
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
         self.conv_tp_weights = TensorProductWeightsBlock(
             num_elements=self.node_attrs_irreps.num_irreps,
@@ -441,20 +393,13 @@ class ResidualElementDependentInteractionBlock(InteractionBlock):
         irreps_mid = irreps_mid.simplify()
         self.irreps_out = linear_out_irreps(irreps_mid, self.target_irreps)
         self.irreps_out = self.irreps_out.simplify()
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.node_feats_irreps,
-            self.node_attrs_irreps,
-            self.irreps_out,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.node_feats_irreps, self.node_attrs_irreps, self.irreps_out
         )
 
     def forward(
@@ -484,27 +429,23 @@ class ResidualElementDependentInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class AgnosticNonlinearInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
             self.node_feats_irreps, self.edge_attrs_irreps, self.target_irreps
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -518,20 +459,13 @@ class AgnosticNonlinearInteractionBlock(InteractionBlock):
         irreps_mid = irreps_mid.simplify()
         self.irreps_out = linear_out_irreps(irreps_mid, self.target_irreps)
         self.irreps_out = self.irreps_out.simplify()
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.irreps_out,
-            self.node_attrs_irreps,
-            self.irreps_out,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.irreps_out, self.node_attrs_irreps, self.irreps_out
         )
 
     def forward(
@@ -561,28 +495,24 @@ class AgnosticNonlinearInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class AgnosticResidualNonlinearInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
             self.node_feats_irreps, self.edge_attrs_irreps, self.target_irreps
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -596,20 +526,13 @@ class AgnosticResidualNonlinearInteractionBlock(InteractionBlock):
         irreps_mid = irreps_mid.simplify()
         self.irreps_out = linear_out_irreps(irreps_mid, self.target_irreps)
         self.irreps_out = self.irreps_out.simplify()
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.node_feats_irreps,
-            self.node_attrs_irreps,
-            self.irreps_out,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.node_feats_irreps, self.node_attrs_irreps, self.irreps_out
         )
 
     def forward(
@@ -640,15 +563,12 @@ class AgnosticResidualNonlinearInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class RealAgnosticInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
@@ -656,14 +576,13 @@ class RealAgnosticInteractionBlock(InteractionBlock):
             self.edge_attrs_irreps,
             self.target_irreps,
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -674,23 +593,17 @@ class RealAgnosticInteractionBlock(InteractionBlock):
         )
 
         # Linear
+        irreps_mid = irreps_mid.simplify()
         self.irreps_out = self.target_irreps
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.irreps_out,
-            self.node_attrs_irreps,
-            self.irreps_out,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.irreps_out, self.node_attrs_irreps, self.irreps_out
         )
-        self.reshape = reshape_irreps(self.irreps_out, cueq_config=self.cueq_config)
+        self.reshape = reshape_irreps(self.irreps_out)
 
     def forward(
         self,
@@ -722,15 +635,12 @@ class RealAgnosticInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class RealAgnosticResidualInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
@@ -738,14 +648,13 @@ class RealAgnosticResidualInteractionBlock(InteractionBlock):
             self.edge_attrs_irreps,
             self.target_irreps,
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -756,23 +665,17 @@ class RealAgnosticResidualInteractionBlock(InteractionBlock):
         )
 
         # Linear
+        irreps_mid = irreps_mid.simplify()
         self.irreps_out = self.target_irreps
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.node_feats_irreps,
-            self.node_attrs_irreps,
-            self.hidden_irreps,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.node_feats_irreps, self.node_attrs_irreps, self.hidden_irreps
         )
-        self.reshape = reshape_irreps(self.irreps_out, cueq_config=self.cueq_config)
+        self.reshape = reshape_irreps(self.irreps_out)
 
     def forward(
         self,
@@ -804,15 +707,12 @@ class RealAgnosticResidualInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class RealAgnosticDensityInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
@@ -820,14 +720,13 @@ class RealAgnosticDensityInteractionBlock(InteractionBlock):
             self.edge_attrs_irreps,
             self.target_irreps,
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -838,22 +737,17 @@ class RealAgnosticDensityInteractionBlock(InteractionBlock):
         )
 
         # Linear
+        irreps_mid = irreps_mid.simplify()
         self.irreps_out = self.target_irreps
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.irreps_out,
-            self.node_attrs_irreps,
-            self.irreps_out,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.irreps_out, self.node_attrs_irreps, self.irreps_out
         )
+        self.reshape = reshape_irreps(self.irreps_out)
 
         # Density normalization
         self.density_fn = nn.FullyConnectedNet(
@@ -864,7 +758,7 @@ class RealAgnosticDensityInteractionBlock(InteractionBlock):
             torch.nn.functional.silu,
         )
         # Reshape
-        self.reshape = reshape_irreps(self.irreps_out, cueq_config=self.cueq_config)
+        self.reshape = reshape_irreps(self.irreps_out)
 
     def forward(
         self,
@@ -900,16 +794,12 @@ class RealAgnosticDensityInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class RealAgnosticDensityResidualInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
-
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
@@ -917,14 +807,13 @@ class RealAgnosticDensityResidualInteractionBlock(InteractionBlock):
             self.edge_attrs_irreps,
             self.target_irreps,
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
@@ -935,22 +824,17 @@ class RealAgnosticDensityResidualInteractionBlock(InteractionBlock):
         )
 
         # Linear
+        irreps_mid = irreps_mid.simplify()
         self.irreps_out = self.target_irreps
-        self.linear = Linear(
-            irreps_mid,
-            self.irreps_out,
-            internal_weights=True,
-            shared_weights=True,
-            cueq_config=self.cueq_config,
+        self.linear = o3.Linear(
+            irreps_mid, self.irreps_out, internal_weights=True, shared_weights=True
         )
 
         # Selector TensorProduct
-        self.skip_tp = FullyConnectedTensorProduct(
-            self.node_feats_irreps,
-            self.node_attrs_irreps,
-            self.hidden_irreps,
-            cueq_config=self.cueq_config,
+        self.skip_tp = o3.FullyConnectedTensorProduct(
+            self.node_feats_irreps, self.node_attrs_irreps, self.hidden_irreps
         )
+        self.reshape = reshape_irreps(self.irreps_out)
 
         # Density normalization
         self.density_fn = nn.FullyConnectedNet(
@@ -962,7 +846,7 @@ class RealAgnosticDensityResidualInteractionBlock(InteractionBlock):
         )
 
         # Reshape
-        self.reshape = reshape_irreps(self.irreps_out, cueq_config=self.cueq_config)
+        self.reshape = reshape_irreps(self.irreps_out)
 
     def forward(
         self,
@@ -998,16 +882,13 @@ class RealAgnosticDensityResidualInteractionBlock(InteractionBlock):
 @compile_mode("script")
 class RealAgnosticAttResidualInteractionBlock(InteractionBlock):
     def _setup(self) -> None:
-        if not hasattr(self, "cueq_config"):
-            self.cueq_config = None
         self.node_feats_down_irreps = o3.Irreps("64x0e")
         # First linear
-        self.linear_up = Linear(
+        self.linear_up = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         # TensorProduct
         irreps_mid, instructions = tp_out_irreps_with_instructions(
@@ -1015,23 +896,21 @@ class RealAgnosticAttResidualInteractionBlock(InteractionBlock):
             self.edge_attrs_irreps,
             self.target_irreps,
         )
-        self.conv_tp = TensorProduct(
+        self.conv_tp = o3.TensorProduct(
             self.node_feats_irreps,
             self.edge_attrs_irreps,
             irreps_mid,
             instructions=instructions,
             shared_weights=False,
             internal_weights=False,
-            cueq_config=self.cueq_config,
         )
 
         # Convolution weights
-        self.linear_down = Linear(
+        self.linear_down = o3.Linear(
             self.node_feats_irreps,
             self.node_feats_down_irreps,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
         input_dim = (
             self.edge_feats_irreps.num_irreps
@@ -1043,21 +922,19 @@ class RealAgnosticAttResidualInteractionBlock(InteractionBlock):
         )
 
         # Linear
+        irreps_mid = irreps_mid.simplify()
         self.irreps_out = self.target_irreps
-        self.linear = Linear(
+        self.linear = o3.Linear(
             irreps_mid,
             self.irreps_out,
             internal_weights=True,
             shared_weights=True,
-            cueq_config=self.cueq_config,
         )
 
-        self.reshape = reshape_irreps(self.irreps_out, cueq_config=self.cueq_config)
+        self.reshape = reshape_irreps(self.irreps_out)
 
         # Skip connection.
-        self.skip_linear = Linear(
-            self.node_feats_irreps, self.hidden_irreps, cueq_config=self.cueq_config
-        )
+        self.skip_linear = o3.Linear(self.node_feats_irreps, self.hidden_irreps)
 
     def forward(
         self,
