+++ mace/modules/irreps_tools.py
@@ -4,13 +4,11 @@
 # This program is distributed under the MIT License (see MIT.md)
 ###########################################################################################
 
-from typing import List, Optional, Tuple
+from typing import List, Tuple
 
 import torch
 from e3nn import o3
 from e3nn.util.jit import compile_mode
-
-from mace.modules.wrapper_ops import CuEquivarianceConfig
 
 
 # Based on mir-group/nequip
@@ -66,12 +64,9 @@ def linear_out_irreps(irreps: o3.Irreps, target_irreps: o3.Irreps) -> o3.Irreps:
 
 @compile_mode("script")
 class reshape_irreps(torch.nn.Module):
-    def __init__(
-        self, irreps: o3.Irreps, cueq_config: Optional[CuEquivarianceConfig] = None
-    ) -> None:
+    def __init__(self, irreps: o3.Irreps) -> None:
         super().__init__()
         self.irreps = o3.Irreps(irreps)
-        self.cueq_config = cueq_config
         self.dims = []
         self.muls = []
         for mul, ir in self.irreps:
@@ -86,19 +81,8 @@ class reshape_irreps(torch.nn.Module):
         for mul, d in zip(self.muls, self.dims):
             field = tensor[:, ix : ix + mul * d]  # [batch, sample, mul * repr]
             ix += mul * d
-            if hasattr(self, "cueq_config") and self.cueq_config is not None:
-                if self.cueq_config.layout_str == "mul_ir":
-                    field = field.reshape(batch, mul, d)
-                else:
-                    field = field.reshape(batch, d, mul)
-            else:
-                field = field.reshape(batch, mul, d)
+            field = field.reshape(batch, mul, d)
             out.append(field)
-
-        if hasattr(self, "cueq_config") and self.cueq_config is not None:
-            if self.cueq_config.layout_str == "mul_ir":
-                return torch.cat(out, dim=-1)
-            return torch.cat(out, dim=-2)
         return torch.cat(out, dim=-1)
