--- mace/modules/wrapper_ops.py
@@ -1,269 +1,0 @@
-"""
-Wrapper class for o3.Linear that optionally uses cuet.Linear
-"""
-
-import dataclasses
-import itertools
-import types
-from typing import Iterator, List, Optional
-
-import numpy as np
-import torch
-from e3nn import o3
-
-from mace.modules.symmetric_contraction import SymmetricContraction
-
-try:
-    import cuequivariance as cue
-    import cuequivariance_torch as cuet
-
-    CUET_AVAILABLE = True
-except ImportError:
-    CUET_AVAILABLE = False
-
-if CUET_AVAILABLE:
-
-    class O3_e3nn(cue.O3):
-        def __mul__(  # pylint: disable=no-self-argument
-            rep1: "O3_e3nn", rep2: "O3_e3nn"
-        ) -> Iterator["O3_e3nn"]:
-            return [O3_e3nn(l=ir.l, p=ir.p) for ir in cue.O3.__mul__(rep1, rep2)]
-
-        @classmethod
-        def clebsch_gordan(
-            cls, rep1: "O3_e3nn", rep2: "O3_e3nn", rep3: "O3_e3nn"
-        ) -> np.ndarray:
-            rep1, rep2, rep3 = cls._from(rep1), cls._from(rep2), cls._from(rep3)
-
-            if rep1.p * rep2.p == rep3.p:
-                return o3.wigner_3j(rep1.l, rep2.l, rep3.l).numpy()[None] * np.sqrt(
-                    rep3.dim
-                )
-            return np.zeros((0, rep1.dim, rep2.dim, rep3.dim))
-
-        def __lt__(  # pylint: disable=no-self-argument
-            rep1: "O3_e3nn", rep2: "O3_e3nn"
-        ) -> bool:
-            rep2 = rep1._from(rep2)
-            return (rep1.l, rep1.p) < (rep2.l, rep2.p)
-
-        @classmethod
-        def iterator(cls) -> Iterator["O3_e3nn"]:
-            for l in itertools.count(0):
-                yield O3_e3nn(l=l, p=1 * (-1) ** l)
-                yield O3_e3nn(l=l, p=-1 * (-1) ** l)
-
-else:
-    print(
-        "cuequivariance or cuequivariance_torch is not available. Cuequivariance acceleration will be disabled."
-    )
-
-
-@dataclasses.dataclass
-class CuEquivarianceConfig:
-    """Configuration for cuequivariance acceleration"""
-
-    enabled: bool = False
-    layout: str = "mul_ir"  # One of: mul_ir, ir_mul
-    layout_str: str = "mul_ir"
-    group: str = "O3"
-    optimize_all: bool = False  # Set to True to enable all optimizations
-    optimize_linear: bool = False
-    optimize_channelwise: bool = False
-    optimize_symmetric: bool = False
-    optimize_fctp: bool = False
-
-    def __post_init__(self):
-        if self.enabled and CUET_AVAILABLE:
-            self.layout_str = self.layout
-            self.layout = getattr(cue, self.layout)
-            self.group = (
-                O3_e3nn if self.group == "O3_e3nn" else getattr(cue, self.group)
-            )
-
-
-class Linear:
-    """Returns either a cuet.Linear or o3.Linear based on config"""
-
-    def __new__(
-        cls,
-        irreps_in: o3.Irreps,
-        irreps_out: o3.Irreps,
-        shared_weights: bool = True,
-        internal_weights: bool = True,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
-        if (
-            CUET_AVAILABLE
-            and cueq_config is not None
-            and cueq_config.enabled
-            and (cueq_config.optimize_all or cueq_config.optimize_linear)
-        ):
-            instance = cuet.Linear(
-                cue.Irreps(cueq_config.group, irreps_in),
-                cue.Irreps(cueq_config.group, irreps_out),
-                layout=cueq_config.layout,
-                shared_weights=shared_weights,
-                optimize_fallback=True,
-            )
-            instance.original_forward = instance.forward
-
-            def cuet_forward(self, x: torch.Tensor) -> torch.Tensor:
-                return self.original_forward(x, use_fallback=True)
-
-            instance.forward = types.MethodType(cuet_forward, instance)
-            return instance
-
-        return o3.Linear(
-            irreps_in,
-            irreps_out,
-            shared_weights=shared_weights,
-            internal_weights=internal_weights,
-        )
-
-
-class TensorProduct:
-    """Wrapper around o3.TensorProduct/cuet.ChannelwiseTensorProduct"""
-
-    def __new__(
-        cls,
-        irreps_in1: o3.Irreps,
-        irreps_in2: o3.Irreps,
-        irreps_out: o3.Irreps,
-        instructions: Optional[List] = None,
-        shared_weights: bool = False,
-        internal_weights: bool = False,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
-        if (
-            CUET_AVAILABLE
-            and cueq_config is not None
-            and cueq_config.enabled
-            and (cueq_config.optimize_all or cueq_config.optimize_channelwise)
-        ):
-            instance = cuet.ChannelWiseTensorProduct(
-                cue.Irreps(cueq_config.group, irreps_in1),
-                cue.Irreps(cueq_config.group, irreps_in2),
-                cue.Irreps(cueq_config.group, irreps_out),
-                layout=cueq_config.layout,
-                shared_weights=shared_weights,
-                internal_weights=internal_weights,
-            )
-            instance.original_forward = instance.forward
-
-            def cuet_forward(
-                self, x: torch.Tensor, y: torch.Tensor, z: torch.Tensor
-            ) -> torch.Tensor:
-                return self.original_forward(x, y, z, use_fallback=None)
-
-            instance.forward = types.MethodType(cuet_forward, instance)
-            return instance
-
-        return o3.TensorProduct(
-            irreps_in1,
-            irreps_in2,
-            irreps_out,
-            instructions=instructions,
-            shared_weights=shared_weights,
-            internal_weights=internal_weights,
-        )
-
-
-class FullyConnectedTensorProduct:
-    """Wrapper around o3.FullyConnectedTensorProduct/cuet.FullyConnectedTensorProduct"""
-
-    def __new__(
-        cls,
-        irreps_in1: o3.Irreps,
-        irreps_in2: o3.Irreps,
-        irreps_out: o3.Irreps,
-        shared_weights: bool = True,
-        internal_weights: bool = True,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
-        if (
-            CUET_AVAILABLE
-            and cueq_config is not None
-            and cueq_config.enabled
-            and (cueq_config.optimize_all or cueq_config.optimize_fctp)
-        ):
-            instance = cuet.FullyConnectedTensorProduct(
-                cue.Irreps(cueq_config.group, irreps_in1),
-                cue.Irreps(cueq_config.group, irreps_in2),
-                cue.Irreps(cueq_config.group, irreps_out),
-                layout=cueq_config.layout,
-                shared_weights=shared_weights,
-                internal_weights=internal_weights,
-                optimize_fallback=True,
-            )
-            instance.original_forward = instance.forward
-
-            def cuet_forward(
-                self, x: torch.Tensor, attrs: torch.Tensor
-            ) -> torch.Tensor:
-                return self.original_forward(x, attrs, use_fallback=True)
-
-            instance.forward = types.MethodType(cuet_forward, instance)
-            return instance
-
-        return o3.FullyConnectedTensorProduct(
-            irreps_in1,
-            irreps_in2,
-            irreps_out,
-            shared_weights=shared_weights,
-            internal_weights=internal_weights,
-        )
-
-
-class SymmetricContractionWrapper:
-    """Wrapper around SymmetricContraction/cuet.SymmetricContraction"""
-
-    def __new__(
-        cls,
-        irreps_in: o3.Irreps,
-        irreps_out: o3.Irreps,
-        correlation: int,
-        num_elements: Optional[int] = None,
-        cueq_config: Optional[CuEquivarianceConfig] = None,
-    ):
-        if (
-            CUET_AVAILABLE
-            and cueq_config is not None
-            and cueq_config.enabled
-            and (cueq_config.optimize_all or cueq_config.optimize_symmetric)
-        ):
-            instance = cuet.SymmetricContraction(
-                cue.Irreps(cueq_config.group, irreps_in),
-                cue.Irreps(cueq_config.group, irreps_out),
-                layout_in=cue.ir_mul,
-                layout_out=cueq_config.layout,
-                contraction_degree=correlation,
-                num_elements=num_elements,
-                original_mace=True,
-                dtype=torch.get_default_dtype(),
-                math_dtype=torch.get_default_dtype(),
-            )
-            instance.original_forward = instance.forward
-            instance.layout = cueq_config.layout
-
-            def cuet_forward(
-                self, x: torch.Tensor, attrs: torch.Tensor
-            ) -> torch.Tensor:
-                if self.layout == cue.mul_ir:
-                    x = torch.transpose(x, 1, 2)
-                index_attrs = torch.nonzero(attrs)[:, 1].int()
-                return self.original_forward(
-                    x.flatten(1),
-                    index_attrs,
-                    use_fallback=None,
-                )
-
-            instance.forward = types.MethodType(cuet_forward, instance)
-            return instance
-
-        return SymmetricContraction(
-            irreps_in=irreps_in,
-            irreps_out=irreps_out,
-            correlation=correlation,
-            num_elements=num_elements,
-        )
