+++ setup.cfg
@@ -29,7 +29,6 @@ install_requires =
     GitPython
     pyYAML
     tqdm
-    cuequivariance-torch
     # for plotting:
     matplotlib
     pandas
@@ -43,10 +42,7 @@ console_scripts =
     mace_run_train  =  mace.cli.run_train:main
     mace_prepare_data  =  mace.cli.preprocess_data:main
     mace_finetuning = mace.cli.fine_tuning_select:main
-    mace_convert_device  =  mace.cli.convert_device:main
-    mace_select_head =  mace.cli.select_head:main
-    mace_e3nn_cueq = mace.cli.convert_e3nn_cueq:main
-    mace_cueq_to_e3nn = mace.cli.convert_cueq_e3nn:main
+    mace_convert_dev  =  mace.cli.convert_dev:main
 
 [options.extras_require]
 wandb = wandb
@@ -60,5 +56,3 @@ dev =
     pytest-benchmark
     pylint
 schedulefree = schedulefree
-cueq-cuda-11 = cuequivariance-ops-torch-cu11
-cueq-cuda-12 = cuequivariance-ops-torch-cu12
