+++ tests/test_foundations.py
@@ -12,7 +12,7 @@ from mace import data, modules, tools
 from mace.calculators import mace_mp, mace_off
 from mace.tools import torch_geometric
 from mace.tools.finetuning_utils import load_foundations_elements
-from mace.tools.scripts_utils import extract_config_mace_model, remove_pt_head
+from mace.tools.scripts_utils import extract_config_mace_model
 from mace.tools.utils import AtomicNumberTable
 
 MODEL_PATH = (
@@ -208,240 +208,3 @@ def test_extract_config(model):
     for key in output.keys():
         if isinstance(output[key], torch.Tensor):
             assert torch.allclose(output[key], output_copy[key], atol=1e-5)
-
-
-def test_remove_pt_head():
-    # Set up test data
-    torch.manual_seed(42)
-    atomic_energies_pt_head = np.array([[1.0, 2.0], [3.0, 4.0]], dtype=float)
-    z_table = AtomicNumberTable([1, 8])  # H and O
-
-    # Create multihead model
-    model_config = {
-        "r_max": 5.0,
-        "num_bessel": 8,
-        "num_polynomial_cutoff": 5,
-        "max_ell": 2,
-        "interaction_cls": modules.interaction_classes[
-            "RealAgnosticResidualInteractionBlock"
-        ],
-        "interaction_cls_first": modules.interaction_classes[
-            "RealAgnosticResidualInteractionBlock"
-        ],
-        "num_interactions": 2,
-        "num_elements": len(z_table),
-        "hidden_irreps": o3.Irreps("32x0e + 32x1o"),
-        "MLP_irreps": o3.Irreps("16x0e"),
-        "gate": torch.nn.functional.silu,
-        "atomic_energies": atomic_energies_pt_head,
-        "avg_num_neighbors": 8,
-        "atomic_numbers": z_table.zs,
-        "correlation": 3,
-        "heads": ["pt_head", "DFT"],
-        "atomic_inter_scale": [1.0, 1.0],
-        "atomic_inter_shift": [0.0, 0.1],
-    }
-
-    model = modules.ScaleShiftMACE(**model_config)
-
-    # Create test molecule
-    mol = molecule("H2O")
-    config_pt_head = data.Configuration(
-        atomic_numbers=mol.numbers,
-        positions=mol.positions,
-        energy=1.0,
-        forces=np.random.randn(len(mol), 3),
-        head="DFT",
-    )
-    atomic_data = data.AtomicData.from_config(
-        config_pt_head, z_table=z_table, cutoff=5.0, heads=["pt_head", "DFT"]
-    )
-    dataloader = torch_geometric.dataloader.DataLoader(
-        dataset=[atomic_data], batch_size=1, shuffle=False
-    )
-    batch = next(iter(dataloader))
-    # Test original mode
-    output_orig = model(batch)
-
-    # Convert to single head model
-    new_model = remove_pt_head(model, head_to_keep="DFT")
-
-    # Basic structure tests
-    assert len(new_model.heads) == 1
-    assert new_model.heads[0] == "DFT"
-    assert new_model.atomic_energies_fn.atomic_energies.shape[0] == 1
-    assert len(torch.atleast_1d(new_model.scale_shift.scale)) == 1
-    assert len(torch.atleast_1d(new_model.scale_shift.shift)) == 1
-
-    # Test output consistency
-    atomic_data = data.AtomicData.from_config(
-        config_pt_head, z_table=z_table, cutoff=5.0, heads=["DFT"]
-    )
-    dataloader = torch_geometric.dataloader.DataLoader(
-        dataset=[atomic_data], batch_size=1, shuffle=False
-    )
-    batch = next(iter(dataloader))
-    output_new = new_model(batch)
-    torch.testing.assert_close(
-        output_orig["energy"], output_new["energy"], rtol=1e-5, atol=1e-5
-    )
-    torch.testing.assert_close(
-        output_orig["forces"], output_new["forces"], rtol=1e-5, atol=1e-5
-    )
-
-
-def test_remove_pt_head_multihead():
-    # Set up test data
-    torch.manual_seed(42)
-    atomic_energies_pt_head = np.array(
-        [
-            [1.0, 2.0],  # H energies for each head
-            [3.0, 4.0],  # O energies for each head
-        ]
-        * 2
-    )
-    z_table = AtomicNumberTable([1, 8])  # H and O
-
-    # Create multihead model
-    model_config = {
-        "r_max": 5.0,
-        "num_bessel": 8,
-        "num_polynomial_cutoff": 5,
-        "max_ell": 2,
-        "interaction_cls": modules.interaction_classes[
-            "RealAgnosticResidualInteractionBlock"
-        ],
-        "interaction_cls_first": modules.interaction_classes[
-            "RealAgnosticResidualInteractionBlock"
-        ],
-        "num_interactions": 2,
-        "num_elements": len(z_table),
-        "hidden_irreps": o3.Irreps("32x0e + 32x1o"),
-        "MLP_irreps": o3.Irreps("16x0e"),
-        "gate": torch.nn.functional.silu,
-        "atomic_energies": atomic_energies_pt_head,
-        "avg_num_neighbors": 8,
-        "atomic_numbers": z_table.zs,
-        "correlation": 3,
-        "heads": ["pt_head", "DFT", "MP2", "CCSD"],
-        "atomic_inter_scale": [1.0, 1.0, 1.0, 1.0],
-        "atomic_inter_shift": [0.0, 0.1, 0.2, 0.3],
-    }
-
-    model = modules.ScaleShiftMACE(**model_config)
-
-    # Create test configurations for each head
-    mol = molecule("H2O")
-    configs = {}
-    atomic_datas = {}
-    dataloaders = {}
-    original_outputs = {}
-
-    # First get outputs from original model for each head
-    for head in model.heads:
-        config_pt_head = data.Configuration(
-            atomic_numbers=mol.numbers,
-            positions=mol.positions,
-            energy=1.0,
-            forces=np.random.randn(len(mol), 3),
-            head=head,
-        )
-        configs[head] = config_pt_head
-
-        atomic_data = data.AtomicData.from_config(
-            config_pt_head, z_table=z_table, cutoff=5.0, heads=model.heads
-        )
-        atomic_datas[head] = atomic_data
-
-        dataloader = torch_geometric.dataloader.DataLoader(
-            dataset=[atomic_data], batch_size=1, shuffle=False
-        )
-        dataloaders[head] = dataloader
-
-        batch = next(iter(dataloader))
-        output = model(batch)
-        original_outputs[head] = output
-
-    # Now test each head separately
-    for i, head in enumerate(model.heads):
-        # Convert to single head model
-        new_model = remove_pt_head(model, head_to_keep=head)
-
-        # Basic structure tests
-        assert len(new_model.heads) == 1, f"Failed for head {head}"
-        assert new_model.heads[0] == head, f"Failed for head {head}"
-        assert (
-            new_model.atomic_energies_fn.atomic_energies.shape[0] == 1
-        ), f"Failed for head {head}"
-        assert (
-            len(torch.atleast_1d(new_model.scale_shift.scale)) == 1
-        ), f"Failed for head {head}"
-        assert (
-            len(torch.atleast_1d(new_model.scale_shift.shift)) == 1
-        ), f"Failed for head {head}"
-
-        # Verify scale and shift values
-        assert torch.allclose(
-            new_model.scale_shift.scale, model.scale_shift.scale[i : i + 1]
-        ), f"Failed for head {head}"
-        assert torch.allclose(
-            new_model.scale_shift.shift, model.scale_shift.shift[i : i + 1]
-        ), f"Failed for head {head}"
-
-        # Test output consistency
-        single_head_data = data.AtomicData.from_config(
-            configs[head], z_table=z_table, cutoff=5.0, heads=[head]
-        )
-        single_head_loader = torch_geometric.dataloader.DataLoader(
-            dataset=[single_head_data], batch_size=1, shuffle=False
-        )
-        batch = next(iter(single_head_loader))
-        new_output = new_model(batch)
-
-        # Compare outputs
-        print(
-            original_outputs[head]["energy"],
-            new_output["energy"],
-        )
-        torch.testing.assert_close(
-            original_outputs[head]["energy"],
-            new_output["energy"],
-            rtol=1e-5,
-            atol=1e-5,
-            msg=f"Energy mismatch for head {head}",
-        )
-        torch.testing.assert_close(
-            original_outputs[head]["forces"],
-            new_output["forces"],
-            rtol=1e-5,
-            atol=1e-5,
-            msg=f"Forces mismatch for head {head}",
-        )
-
-    # Test error cases
-    with pytest.raises(ValueError, match="Head non_existent not found in model"):
-        remove_pt_head(model, head_to_keep="non_existent")
-
-    # Test default behavior (first non-PT head)
-    default_model = remove_pt_head(model)
-    assert default_model.heads[0] == "DFT"
-
-    # Additional test: check if each model's computation graph is independent
-    models = {head: remove_pt_head(model, head_to_keep=head) for head in model.heads}
-    results = {}
-
-    for head, head_model in models.items():
-        single_head_data = data.AtomicData.from_config(
-            configs[head], z_table=z_table, cutoff=5.0, heads=[head]
-        )
-        single_head_loader = torch_geometric.dataloader.DataLoader(
-            dataset=[single_head_data], batch_size=1, shuffle=False
-        )
-        batch = next(iter(single_head_loader))
-        results[head] = head_model(batch)
-
-    # Verify each model produces different outputs
-    energies = torch.stack([results[head]["energy"] for head in model.heads])
-    assert not torch.allclose(
-        energies[0], energies[1], rtol=1e-3
-    ), "Different heads should produce different outputs"
